requirements.txt
asn1crypto==1.5.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==40.0.2
filelock==3.11.0
idna==3.4
numpy==1.24.2
oscrypto==1.3.0
packaging==23.1
pandas==1.5.3
psycopg2==2.9.6
pyarrow==10.0.1
pycparser==2.21
pycryptodomex==3.17
PyJWT==2.6.0
pyOpenSSL==23.1.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
six==1.16.0
snowflake-connector-python==3.0.2
snowflake-sqlalchemy==1.4.7
SQLAlchemy==1.4.47
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
